+++ src/decode/common.h
@@ -140,17 +140,15 @@ typedef struct CRCscan {
 typedef struct ProbedPage {
    int64_t page_start, page_end;
    int64_t after_previous_page_start;
-   // FIXME: 64-bit sample positions
-   uint32_t first_decoded_sample;
-   uint32_t last_decoded_sample;
+   uint64_t first_decoded_sample;
+   uint64_t last_decoded_sample;
 } ProbedPage;
 
 struct stb_vorbis {
     /* Basic stream information. */
     unsigned int sample_rate;
     int channels;
-    // FIXME: 64-bit sample positions
-    uint32_t total_samples;
+    uint64_t total_samples;
     int64_t stream_len;  // from open()
 
     /* Callbacks for stream reading.  The seek and tell callbacks are only
@@ -194,8 +192,7 @@ struct stb_vorbis {
    uint8_t ***part_classdata;
 #endif
 
-   // FIXME: 64-bit sample positions
-   uint32_t current_loc; // sample location of next frame to decode
+   uint64_t current_loc; // sample location of next frame to decode
    bool current_loc_valid;
 
   // temporary buffer for IMDCT
@@ -238,8 +235,7 @@ struct stb_vorbis {
    ProbedPage p_first, p_last;
 
    int end_seg_with_known_loc;
-   // FIXME: 64-bit sample positions
-   uint32_t known_loc_for_packet;
+   uint64_t known_loc_for_packet;
    int discard_samples_deferred;
 };
